backend/requirements.txt
flask
flask-cors
pandas
numpy
networkx
sentence-transformers
faiss-cpu
nltk
python-dotenv
node2vec
tqdm
langchain
langchain-openai
neo4j
flask-socketio
gevent
googletrans
langdetect
pyspellchecker